cognito/userpool-appclient-callback/assets/get-user/requirements.txt
python-jose < 3.2.0